requirements.txt
asgiref==3.2.10
boto3==1.14.31
botocore==1.17.31
Django==3.0.8
django-ckeditor==5.9.0
django-crispy-forms==1.9.2
django-filter==2.3.0
django-js-asset==1.2.2
django-storages==1.9.1
docutils==0.15.2
gunicorn==20.0.4
jmespath==0.10.0
Pillow
psycopg2
python-dateutil
pytz
s3transfer
six
sqlparse
urllib3
whitenoise

